pyproject.toml
[tool.poetry]
name = "test-referalsystemservice"
version = "0.1.0"
description = ""
authors = ["NormanwOw <normjkeee@vk.com>"]
readme = "README.md"
packages = [{include = "test"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
flake8 = "^7.1.1"
flake8-print = "^5.0.0"
pre-commit = "^4.1.0"
pydantic-settings = "^2.7.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
passlib = "^1.7.4"
bcrypt = "4.0.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
pytz = "^2025.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
